Migrate most code to non-main package
