+++ .gitignore
@@ -1,4 +1,4 @@
-gron
+/gron
 *.tgz
 *.zip
 *.swp
