+++ internal/gron/identifier_test.go
@@ -1,4 +1,4 @@
-package main
+package gron
 
 import "testing"
