+++ internal/gron/statements_test.go
@@ -1,4 +1,4 @@
-package main
+package gron
 
 import (
 	"bytes"
