+++ internal/gron/testdata/large-line.gron
