+++ internal/gron/testdata/long-stream.gron
