+++ internal/gron/testdata/scalar-stream.gron
