+++ internal/gron/testdata/scalar-stream.jgron
