+++ internal/gron/testdata/scalar-stream.json
