+++ internal/gron/testdata/stream.gron
