+++ internal/gron/testdata/stream.jgron
