+++ internal/gron/testdata/stream.json
