+++ internal/gron/ungron_test.go
@@ -1,4 +1,4 @@
-package main
+package gron
 
 import (
 	"reflect"
